ref: Move Credentials Button to a separate file
